Add source code walkthrough as code comment
Add source code walkthrough as code comment to main.go. This will be
of use for keeping the code structured, both for myself and for
contributors.

In the event that the program structure changes, I will reflect
those changes accordingly. As of now, this is the primary source
for understanding the source code.